requirements.txt
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
openai>=1.0.0
langchain>=0.0.300
python-dotenv>=1.0.0
loguru>=0.7.0
fastapi>=0.100.0
uvicorn>=0.23.0
tabulate>=0.9.0
rich>=13.0.0
psycopg2-binary>=2.9.5
anthropic>=0.18.0
google-generativeai>=0.3.0
requests>=2.31.0
scikit-learn>=1.2.0
numpy>=1.21.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4  # 使用 GPU 版本請改用 faiss-gpu
joblib>=1.2.0
# 視覺化支持
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
# 測試工具
pytest>=7.0.0
pytest-cov>=4.0.0